fix blog label; refactor section sub-element names
