+++ shared/components/card/_article.scss
@@ -57,10 +57,12 @@
 
 .liveblog__badge {
 	@include oTypographySansBold(s);
+	float: left;
+	line-height: 21px;
 	display: inline-block;
 	text-transform: uppercase;
-	padding: 0.1em 0.5em 0.3em;
-	margin: 0 0.4em 0 0;
+	padding: 0 5px;
+	margin-right: 5px;
 	color: white;
 }
