+++ shared/components/section/_section.scss
@@ -1,6 +1,6 @@
 @import "o-forms/main";
-@import "./section-meta/section-meta";
-@import "./section-sources/section-sources";
+@import "meta/meta";
+@import "sources/sources";
 @import "../ad/ad";
 @import "../card/card";
 @import "../myft-promo/myft-promo";
