+++ shared/components/section/content/content.js
