+++ shared/components/section/meta/_meta.scss
