+++ shared/components/section/sources/_sources.scss
