+++ shared/components/section/sources/sources.js
